+++ src/Bootstrap/Services.php
@@ -274,6 +274,7 @@ class Services implements ServicesInterface
 		});
 
 		// Active gateway service
+		// @deprecated call `gateways` instead and select the appropriate gateway
 		$services['gateway'] = function($c) {
 			$gateway = $c['cfg']->payment->gateway;
 
@@ -301,6 +302,14 @@ class Services implements ServicesInterface
 			$gateways->setSort(null);
 
 			return $gateways;
+		};
+
+		$services['payment.gateway.loader'] = function ($c) {
+			return new Ecommerce\Payment\PaymentGatewayRecordLoader($c['payment.loader'], $c['gateway.collection'], $c['db.query.builder.factory']);
+		};
+
+		$services['payment.gateway.edit'] = function ($c) {
+			return new Ecommerce\Payment\PaymentGatewayRecordEdit($c['db.query']);
 		};
 	}
 }