+++ src/EventListener/OrderListener.php
@@ -27,9 +27,18 @@ class OrderListener extends BaseListener implements SubscriberInterface
 			Order\Events::UPDATE_FAILED => [
 				['redirectToHome']
 			],
+			Order\Events::ORDER_CANCEL_REFUND => [
+				['setOrderRefundController']
+			],
+			Order\Events::ITEM_CANCEL_REFUND => [
+				['setItemRefundController']
+			],
 		);
 	}
 
+	/**
+	 * @param Order\Event\Event $event
+	 */
 	public function sendOrderConfirmationMail(Order\Event\Event $event)
 	{
 		$order = $event->getOrder();
@@ -46,6 +55,9 @@ class OrderListener extends BaseListener implements SubscriberInterface
 		}
 	}
 
+	/**
+	 * @param Order\Event\UpdateFailedEvent $event
+	 */
 	public function redirectToHome(Order\Event\UpdateFailedEvent $event)
 	{
 		$page = $this->get('cms.page.loader')->getHomepage();
@@ -54,5 +66,62 @@ class OrderListener extends BaseListener implements SubscriberInterface
 		$redirectEvent->setResponse(new RedirectResponse($page->slug));
 
 		$this->get('event.dispatcher')->dispatch(Page\Event\Event::RENDER_SET_RESPONSE, $redirectEvent);
+	}
+
+	/**
+	 * @param Order\Event\CancelEvent $event
+	 */
+	public function setOrderRefundController(Order\Event\CancelEvent $event)
+	{
+		$this->_setRefundController($event, 'order');
+	}
+
+	/**
+	 * @param Order\Event\CancelEvent $event
+	 */
+	public function setItemRefundController(Order\Event\CancelEvent $event)
+	{
+		$this->_setRefundController($event, 'item');
+	}
+
+	/**
+	 * Set controller and parameters to process refund via gateway. Can configure cancellation for
+	 * orders or individual items by giving 'order' or 'item' as the second parameter
+	 *
+	 * @param Order\Event\CancelEvent $event
+	 * @param $type
+	 * @throws \LogicException                  Throws exception if there are no payments to refund
+	 */
+	private function _setRefundController(Order\Event\CancelEvent $event, $type)
+	{
+		$types = ['order', 'item'];
+		if (!in_array($type, $types)) {
+			throw new \LogicException('Invalid refund type, must be in array: ' . implode(', ', $types));
+		}
+
+		$paymentReference = null;
+		$gateway = null;
+
+		foreach ($event->getOrder()->payments as $payment) {
+			$gateway = $this->get('payment.gateway.loader')->getGatewayByPayment($payment->payment);
+			$paymentReference = $payment->reference;
+			break;
+		}
+
+		if (!$gateway) {
+			throw new \LogicException('Could not load gateway, no payments to refund');
+		}
+
+		$controller = 'Message:Mothership:Commerce::Controller:Order:Cancel:Refund';
+
+		$event->setControllerReference($gateway->getRefundControllerReference());
+		$event->setParams([
+			'payable' => $event->getRefund(),
+			'reference' => $paymentReference,
+			'stages' => [
+				'failure' => $controller . '#' . $type . 'Failure',
+				'success' => $controller . '#' . $type . 'Success',
+			],
+		]);
 	}
 }