+++ src/Payment/GatewayNotFoundException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Message\Mothership\Ecommerce\Payment;
+
+/**
+ * @author Samuel Trangmar-Keates <sam@message.co.uk>
+ *
+ * Exception thrown if no gateway record is found
+ */
+class GatewayNotFoundException extends \LogicExceptiion
+{}